Extract payment gateways logic into .Gateways project
