+++ RubyRemit.Gateways/Services/PremiumPaymentService.cs
@@ -1,9 +1,9 @@
 ﻿using RubyRemit.Infrastructure.PaymentGateways.Contracts;
 using System;
 
-namespace RubyRemit.Infrastructure.PaymentGateways.Services
+namespace RubyRemit.Gateways.Services
 {
- 
+
     public class PremiumPaymentService : PaymentServiceBase, IExpensivePaymentGateway
     {
         public PremiumPaymentService(decimal? commissionRate)
