+++ RubyRemit.Infrastructure/PaymentGateways/Contracts/IPaymentGateway.cs
@@ -1,6 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
+﻿using RubyRemit.Domain.DTOs;
 using System.Threading.Tasks;
 
 namespace RubyRemit.Infrastructure.PaymentGateways.Contracts
@@ -9,10 +7,13 @@ namespace RubyRemit.Infrastructure.PaymentGateways.Contracts
     {        
         public string ServiceName { get; }
 
+
         public decimal CommissionRate { get; }
+
 
         public decimal CalculateCommission(decimal amount, decimal commissionRate);
 
-        public Task<bool> ProcessTransaction(string cardNo, string holder, DateTime expDate, string secCode, decimal transAmt, out string message);
-    }
+
+        public Task<GatewayResponse> ProcessTransaction(PaymentRequestBody request);
+   }
 }
